% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildPandF}
\alias{buildPandF}
\title{Build P and F}
\usage{
buildPandF(
  Fdat = fecundity,
  cdata = catch_data,
  Fproj = Fproj,
  Pdat = Pdat,
  Pper = Pper,
  return.periods = T
)
}
\arguments{
\item{Fdat}{Observed fecundity data}

\item{cdata}{Catch data}

\item{Fproj}{Type of projection}

\item{Pdat}{Birth ogive data}

\item{Pper}{Time periods for the various birth ogive curves}

\item{return.periods}{Set True if to return the various periods}
}
\value{
Fdt Vector of fecundity data

Pmat Matrix of birth ogive data

Pper Return the Pper variable
}
\description{
Create the birth ogive matrix and the fecundity rate vector
}
\examples{


}
\keyword{Fecundity}
