% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{eq.quota.helper.N1}
\alias{eq.quota.helper.N1}
\title{Internal helper function for finding equilibrium quota}
\usage{
eq.quota.helper.N1(Tot, population, quota,
  root.dir = "C:/Users/a5406/Documents/Populasjonsmodellering/rSPAMM/rSPAMM-master")
}
\arguments{
\item{Tot}{Total annual catch for population projections}

\item{population}{Choose which population to run the model on (harpwest,harpeast,hooded).}

\item{quota}{Proportional catch of 0 and 1+ animals}

\item{root.dir}{Root directory for population model files}
}
\value{
minN Minimum found for difference between current and projected 1+ population size
}
\description{
Finding N-based equilibrium quota
}
\keyword{model}
\keyword{population}
