% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{find.eq.quota}
\alias{find.eq.quota}
\title{Main function for finding equilibrium quota}
\usage{
find.eq.quota(
  MIN = 1000,
  MAX = 40000,
  quota = c(0, 1),
  data = data,
  parameters = parameters,
  method = "Dbased"
)
}
\arguments{
\item{MIN}{Minimum extent of search window for total catch}

\item{MAX}{Maximum extent of search window for total catch}

\item{quota}{Proportional catch of 0 and 1+ animals}

\item{data}{The data to be analyzed}

\item{parameters}{Parameters used for the model}

\item{method}{Set whether D-based (depletion coefficient D) or N-based (1+ population size) criterion should be used for optimisation  (Dbased/Nbased)}
}
\value{
qEq Optimum quota for achieving equilibrium projected population size
}
\description{
Finding D-based or N-based equilibrium quota
}
\keyword{model}
\keyword{population}
