% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ilogit}
\alias{ilogit}
\title{Inverse logit transformation - type 1}
\usage{
ilogit(x)
}
\arguments{
\item{x}{Number to be inverse logit transformed}
}
\value{
ilogit_x Inverse logit transformed value of x
}
\description{
Performs an inverse logit transforms the input number
}
\examples{
ilogit(x)
}
\keyword{logit}
