% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ilogit1}
\alias{ilogit1}
\title{Inverse logit transformation - type 2}
\usage{
ilogit1(x)
}
\arguments{
\item{x}{Number to be inverse logit transformed}
}
\value{
ilogit_x Inverse logit transformed value of x
}
\description{
Performs an inverse logit transforms the input number
}
\examples{
ilogit1(x) 
}
\keyword{logit}
