% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment_models.R
\name{load.model.object}
\alias{load.model.object}
\title{Load the population model for harp seals and hooded seals}
\usage{
load.model.object(
  dat = data,
  par = parameters,
  template = "harps_and_hoods_population_model2"
)
}
\arguments{
\item{population}{Choose which population to run the model on (harpwest,harpeast,hooded).}

\item{Amax}{Maximum age group. Default is 20 years.}

\item{years_of_prediction}{Number of years in the future to project the model. Default is 15 years.}

\item{Fproj}{Which fecundity rate to use in future projections. Fproj = "mean" uses mean value of observed fecundity rates. Otherwise a fixed Fproj can be set.}
}
\value{
obj Model object to be optimized.
}
\description{
Load the model object to be optimized.
}
\examples{
load.model.object(population = "harpeast")
}
\keyword{model}
\keyword{population}
