% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{Number to be logit transformed}
}
\value{
logit_x Logit transformed value of x
}
\description{
Logit transforms the input number
}
\examples{
logit(x) 
}
\keyword{logit}
