% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment_models.R
\name{model.results}
\alias{model.results}
\title{Get results from optimized population model for harp seals and hooded seals}
\usage{
model.results(dat = data, object = obj, optimized = opt)
}
\arguments{
\item{dat}{Data object that was used to fit the model}

\item{object}{The TMB model object}

\item{optimized}{Output from the optimization}

\item{to.file}{Logical, whether to return results to the workspace (default) or save to file. NOT YET IMPLEMENTED!}
}
\value{
results Results returned to the workspace or saved to file.
}
\description{
Get model results.
}
\examples{
load.model.object(population = "harpeast")
}
\keyword{model}
\keyword{population}
