% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{par.table}
\alias{par.table}
\title{Create table with key parameters from fitted population model}
\usage{
par.table(results = res, dat = data, tab2flex = FALSE)
}
\arguments{
\item{results}{Output from fitted model}

\item{dat}{Original data on estimated pup production (set to NA if these are not to be included).}
}
\value{
table Returns a table with key parameters from fitted population model.
}
\description{
Parameter table
}
\examples{
par.table()
}
\keyword{model}
\keyword{population}
