% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plot.fecundity}
\alias{plot.fecundity}
\title{Plot fecundity data}
\usage{
\method{plot}{fecundity}(results = res, dat = data,
  component = c("N0", "N1"), xLim = NA, yLim = NA,
  plot.legend = TRUE, plot.Nlims = TRUE, projections = TRUE,
  conf.int = TRUE)
}
\arguments{
\item{results}{Output from fitted model}

\item{dat}{Original data on estimated pup production (set to NA if these are not to be included).}

\item{component}{Which population component to plot. Can be either 'N0' for pups, 'N1' for adults, or a vector with both.}

\item{xLim}{Manually set the x axis extent, overrides the default which is the extent of the plotted data.}

\item{yLim}{Manually set the y axis extent, overrides the default which is the extent of the plotted data.}

\item{plot.legend}{Add legend to plot (TRUE/FALSE)}

\item{plot.Nlims}{True/False}

\item{projections}{Plot projections (TRUE/FALSE)}

\item{mean.proj}{True/False}
}
\value{
plot Returns a plot of predicted population size for different population components
}
\description{
Plot the fecundity data used in the model fit.
}
\examples{
plot.N(res, data)
}
\keyword{model}
\keyword{population}
