% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plot.N}
\alias{plot.N}
\title{Plot estimated population sizes from optimized population model for harp seals and hooded seals}
\usage{
\method{plot}{N}(results = res, dat = data, component = c("N0",
  "N1"), xLim = NA, yLim = NA, plot.legend = T, plot.Nlims = T,
  projections = T, mean.proj = F)
}
\arguments{
\item{results}{Output from fitted model}

\item{dat}{Original data on estimated pup production (set to NA if these are not to be included).}

\item{component}{Which opulation component to plot. Can be either 'N0' for pups, 'N1' for adults, or a vector with both.}

\item{xLim}{Manually set the x axis extent, overrides the default which is the extent of the plotted data.}

\item{yLim}{Manually set the y axis extent, overrides the default which is the extent of the plotted data.}
}
\value{
plot Returns a plot of predicted population size for different population components
}
\description{
Plot population trajectories.
}
\examples{
plot.N(res, data)
}
\keyword{model}
\keyword{population}
