% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plot.Pmat}
\alias{plot.Pmat}
\title{Plot birth ogive curves}
\usage{
\method{plot}{Pmat}(dat = data, highlight.last = TRUE)
}
\arguments{
\item{dat}{input data for the model}

\item{highlight.last}{Highlist last observed birth ogive curve (TRUE/FALSE)}
}
\description{
Plot the birth ogive curves for various time periods.
}
\examples{
plot.Pmat()
}
\keyword{birth}
\keyword{ogive}
