% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plot.res.ggplot}
\alias{plot.res.ggplot}
\title{Plot modelled population dynamics}
\usage{
\method{plot}{res.ggplot}(res = res, dat = data, component = c("N0",
  "N1"), xLim = NA, yLim = NA, plot.legend = TRUE,
  plot.Nlims = TRUE, col = c("royalblue", "darkred", "steelblue"),
  projections = TRUE, conf.int = TRUE, add.title = TRUE,
  width = 13, height = 7)
}
\arguments{
\item{dat}{Original data on estimated pup production (set to NA if these are not to be included).}

\item{component}{Which population component to plot. Can be either 'N0' for pups, 'N1' for adults, or a vector with both.}

\item{xLim}{Manually set the x axis extent, overrides the default which is the extent of the plotted data.}

\item{yLim}{Manually set the y axis extent, overrides the default which is the extent of the plotted data.}

\item{plot.legend}{Add legend to plot (TRUE/FALSE)}

\item{plot.Nlims}{Plot horizontal lines indicating 30%, 50% and 70% of 1+ population True/False}

\item{col}{Specify colors used for (pups,1+,error bars)}

\item{projections}{Plot projections (TRUE/FALSE)}

\item{add.title}{Add title to figure (TRUE/FALSE)}

\item{results}{Output from fitted model}
}
\value{
plot Returns a plot of predicted population size for different population components
}
\description{
Plot population trajectories. This function does not work properly, use plot.N() instead.
}
\examples{
plot.res(res, data)
}
\keyword{model}
\keyword{population}
