% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plotFecundity}
\alias{plotFecundity}
\title{Plot fecundity data}
\usage{
plotFecundity(dat = data, include.observations = TRUE, grDev = FALSE)
}
\arguments{
\item{include.observations}{Plot the observed fecundity rates with 95 percent confidence intervals (default = TRUE)}

\item{data}{Data object used in model fit}
}
\value{
plot Returns a plot of predicted population size for different population components
}
\description{
Plot the fecundity data used in the model fit.
}
\examples{
plotFecundity(res, data)
}
\keyword{data}
\keyword{fecundity}
