% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plotOgive}
\alias{plotOgive}
\title{Plot birth ogive curves}
\usage{
plotOgive(dat = data, highlight.last = TRUE)
}
\arguments{
\item{dat}{input data for the model}

\item{highlight.last}{Highlist last observed birth ogive curve (TRUE/FALSE)}
}
\description{
Plot the birth ogive curves for various time periods.
}
\examples{
plotOgive()
}
\keyword{birth}
\keyword{ogive}
