% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plotRes}
\alias{plotRes}
\title{Plot estimated population sizes from optimized population model for harp seals and hooded seals}
\usage{
plotRes(
  results = res,
  data = data,
  component = c("N0", "Ntot"),
  plotNlims = TRUE,
  plotProjections = TRUE,
  plotProjMean = TRUE,
  width = 15,
  height = 10,
  grDev = TRUE
)
}
\arguments{
\item{results}{Output from fitted model}

\item{data}{Original data on estimated pup production (set to NA if these are not to be included).}

\item{component}{Which population component to plot. Can be either 'N0' for pups, 'N1' for adults, 'Ntot' for total abundance or a vector with both. Default showing pups and total abundance.}

\item{plotNlims}{Logical parameter to turn on and off Nlim, N50, and N70 lines, default True}

\item{plotProjections}{Logical parameter to turn on and off plotting of projections, default True}

\item{width}{Width of the figure}

\item{height}{Height of the figure}

\item{grDev}{Logical parameter to decide wether to open a OS independent graphical window}

\item{plotMeanProj}{Logical parameter to turn on and off plotting of the mean projections, default TRUE}
}
\value{
plot Returns a plot of predicted population size for different population components
}
\description{
Plot population trajectories.
}
\examples{
plotRes(res, data)
}
\keyword{model,}
\keyword{plot}
\keyword{population}
\keyword{results}
