% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment_models.R
\name{run.model}
\alias{run.model}
\title{Run the loaded population model for harp seals and hooded seals}
\usage{
run.model(data = data, par = parameters, print.to.screen = TRUE)
}
\arguments{
\item{par}{The initial values of the parameters to be estimated.}

\item{dat}{The data used for optimizing the model.}

\item{print2screen}{Logical parameter to decide wether to output results to screen.}
}
\value{
Output from optimized model.
}
\description{
Optimize the model object.
}
\examples{
run.model()
}
\keyword{model}
\keyword{population}
