% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{eq.quota.helper.D}
\alias{eq.quota.helper.D}
\title{Internal helper function for finding equilibrium quota}
\usage{
eq.quota.helper.D(Tot, dataD, parametersD, quota = c(0, 1))
}
\arguments{
\item{Tot}{Total annual catch for population projections}

\item{dataD}{The data to be analyzed}

\item{parametersD}{Parameters used for the model}

\item{quota}{Proportional catch of 0 and 1+ animals}
}
\value{
minD Minimum found for difference between the depletion coefficient D and 1
}
\description{
Finding D-based equilibrium quota
}
\keyword{model}
\keyword{population}
