% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.options.R
\name{find.N70.quota}
\alias{find.N70.quota}
\title{Main function for finding N70 quota}
\usage{
find.N70.quota(
  MIN = 5000,
  MAX = 50000,
  quota = c(0, 1),
  data = data,
  parameters = parameters,
  method = "Nbased"
)
}
\arguments{
\item{MIN}{Minimum extent of search window for total catch}

\item{MAX}{Maximum extent of search window for total catch}

\item{quota}{Proportional catch of 0 and 1+ animals}

\item{data}{The data to be analyzed}

\item{parameters}{Parameters used for the model}

\item{method}{Set whether D-based (depletion coefficient D) or N-based (total population size) criterion should be used for optimisation (Dbased,Nbased)}
}
\value{
q70 Optimum quota for achieving projected size of 70% of maximum population size
}
\description{
Finding D-based or N-based N70 quota
}
\keyword{model}
\keyword{population}
