% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment_models.R
\name{model.results}
\alias{model.results}
\title{Get results from optimized population model for harp seals and hooded seals}
\usage{
model.results(data, optobject)
}
\arguments{
\item{data}{Data object that was used to fit the model}

\item{optobject}{The TMB model object and the output from the optimization (a list)}
}
\value{
results Results returned to the workspace or saved to file.
}
\description{
Get model results.
}
\examples{
load.model.object(population = "harpeast")
}
\keyword{model}
\keyword{population}
