% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plot.N}
\alias{plot.N}
\title{Plot estimated population sizes from optimized population model for harp seals and hooded seals}
\usage{
\method{plot}{N}(
  results = res,
  data = data,
  component = c("N0", "N1"),
  xLim = NA,
  yLim = NA,
  plot.legend = TRUE,
  plot.Nlims = TRUE,
  projections = TRUE,
  mean.proj = FALSE,
  width = 13,
  height = 7,
  conf.int = TRUE
)
}
\arguments{
\item{results}{Output from fitted model}

\item{data}{Original data on estimated pup production (set to NA if these are not to be included).}

\item{component}{Which population component to plot. Can be either 'N0' for pups, 'N1' for adults, or a vector with both.}

\item{xLim}{Manually set the x axis extent, overrides the default which is the extent of the plotted data.}

\item{yLim}{Manually set the y axis extent, overrides the default which is the extent of the plotted data.}

\item{plot.legend}{Add legend to plot (TRUE/FALSE)}

\item{plot.Nlims}{True/False}

\item{projections}{Plot projections (TRUE/FALSE)}

\item{mean.proj}{True/False}

\item{width}{Width of the figure}

\item{height}{Height of the figure}
}
\value{
plot Returns a plot of predicted population size for different population components
}
\description{
Plot population trajectories.
}
\examples{
plot.N(res, data)
}
\keyword{model}
\keyword{population}
