% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment_models.R
\name{run.model}
\alias{run.model}
\title{Run the loaded population model for harp seals and hooded seals}
\usage{
run.model(object = obj, print.to.screen = TRUE)
}
\arguments{
\item{obj}{Model object to be optimized.}
}
\value{
Output from optimized model.
}
\description{
Optimize the model object.
}
\examples{
run.model()
}
\keyword{model}
\keyword{population}
